applyHistory
