+++ source/js/buyers-guide/search.js
@@ -1,8 +1,5 @@
-/**
- * Set up PNI search functionality, as well as the PNI "ding" toggle.
- */
+import { SearchFilter as SearchFilterObject } from "./search/search-filter.js";
 
-// static values used throughout this code
 const ALL_PRODUCTS = document.querySelectorAll(`figure.product-box`);
 const NO_RESULTS_NOTICE = document.getElementById(
   `product-filter-no-results-notice`
@@ -17,427 +14,11 @@ const toggle = document.querySelector(`#product-filter-pni-toggle`);
 const subcategories = document.querySelectorAll(`.subcategories`);
 const subContainer = document.querySelector(`.subcategory-header`);
 
+
 // TODO: turn this into a static class rather than plain JS object.
 const SearchFilter = {
   init: () => {
-    let pos = { left: 0, x: 0 };
-    const subClasses = subContainer.classList;
-
-    const markScrollStart = (event) => {
-      event.preventDefault();
-      event.stopImmediatePropagation();
-      subClasses.add("cursor-grabbing", "select-none");
-
-      pos = {
-        left: subContainer.scrollLeft,
-        x: event.clientX,
-      };
-
-      [`mousemove`, `touchmove`].forEach((type) =>
-        document.addEventListener(type, markScrollMove)
-      );
-
-      [`mouseup`, `touchend`, `touchcancel`].forEach((type) =>
-        document.addEventListener(type, markScrollEnd)
-      );
-    };
-
-    const markScrollMove = (event) => {
-      subcategories.forEach((subcategory) => {
-        subcategory.classList.add("pointer-events-none");
-      });
-      const dx = event.clientX - pos.x;
-      subContainer.scrollLeft = pos.left - dx;
-    };
-
-    const markScrollEnd = (event) => {
-      event.preventDefault();
-      event.stopImmediatePropagation();
-
-      subcategories.forEach((subcategory) => {
-        subcategory.classList.remove("pointer-events-none");
-      });
-
-      subClasses.remove("cursor-grabbing", "select-none");
-
-      [`mousemove`, `touchmove`].forEach((type) =>
-        document.removeEventListener(type, markScrollMove)
-      );
-
-      [`mouseup`, `touchend`, `touchcancel`].forEach((type) =>
-        document.removeEventListener(type, markScrollEnd)
-      );
-    };
-
-    [`mousedown`, `touchstart`].forEach((type) =>
-      subContainer.addEventListener(type, markScrollStart)
-    );
-
-    const searchBar = document.querySelector(`#product-filter-search`);
-
-    if (!searchBar) {
-      return console.warn(
-        `Could not find the PNI search bar. Search will not be available.`
-      );
-    }
-
-    const searchInput = (SearchFilter.searchInput =
-      searchBar.querySelector(`input`));
-
-    searchInput.addEventListener(`input`, (evt) => {
-      const searchText = searchInput.value.trim();
-
-      if (searchText) {
-        searchBar.classList.add(`has-content`);
-        SearchFilter.filter(searchText);
-      } else {
-        clearText();
-        applyHistory();
-      }
-    });
-
-    const clear = searchBar.querySelector(`.clear-icon`);
-    if (!clear) {
-      return console.warn(
-        `Could not find the PNI search input clear icon. Search will work, but clearing will not.`
-      );
-    }
-
-    const clearText = () => {
-      searchBar.classList.remove(`has-content`);
-      searchInput.value = ``;
-      ALL_PRODUCTS.forEach((product) => {
-        product.classList.remove(`d-none`);
-        product.classList.add(`d-flex`);
-      });
-
-      history.replaceState(
-        {
-          ...history.state,
-          search: "",
-        },
-        SearchFilter.getTitle(categoryTitle.value.trim()),
-        location.href
-      );
-
-      SearchFilter.sortOnCreepiness();
-      SearchFilter.moveCreepyFace();
-    };
-
-    const applyHistory = () => {
-      const { category, parent } = history.state;
-
-      categoryTitle.value = category;
-      parentTitle.value = parent;
-
-      if (parent) {
-        SearchFilter.highlightParent();
-        SearchFilter.toggleSubcategory();
-      } else {
-        document
-          .querySelector(`#multipage-nav a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector(`#pni-nav-mobile a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector(`#multipage-nav a[data-name="${category}"]`)
-          .classList.add(`active`);
-
-        document
-          .querySelector(`#pni-nav-mobile a[data-name="${category}"]`)
-          .classList.add(`active`);
-
-        SearchFilter.toggleSubcategory(true);
-      }
-      SearchFilter.filterCategory(category);
-      SearchFilter.filterSubcategory(parent || category);
-      SearchFilter.updateHeader(category, parent);
-      SearchFilter.sortOnCreepiness();
-      SearchFilter.moveCreepyFace();
-
-      if (history.state?.parent && history.state?.category) {
-        document
-          .querySelector(
-            `a.subcategories[data-name="${history.state?.category}"]`
-          )
-          .scrollIntoView({
-            behavior: "smooth",
-            block: "nearest",
-            inline: "start",
-          });
-      }
-    };
-
-    clear.addEventListener(`click`, (evt) => {
-      evt.preventDefault();
-      searchInput.focus();
-      clearText();
-      applyHistory();
-    });
-
-    const navLinks = document.querySelectorAll(
-      `#multipage-nav a,.category-header,#pni-nav-mobile a`
-    );
-
-    for (const nav of navLinks) {
-      nav.addEventListener("click", (evt) => {
-        evt.stopPropagation();
-
-        if (evt.shiftKey || evt.metaKey || evt.ctrlKey || evt.altKey) {
-          return;
-        }
-
-        evt.preventDefault();
-
-        document
-          .querySelector(`#multipage-nav a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector(`#pni-nav-mobile a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector("#pni-nav-mobile .dropdown-nav")
-          .classList.remove("dropdown-nav-open");
-
-        if (evt.target.dataset.name) {
-          document
-            .querySelector(
-              `#multipage-nav a[data-name="${evt.target.dataset.name}"]`
-            )
-            .classList.add(`active`);
-
-          document
-            .querySelector(
-              `#pni-nav-mobile a[data-name="${evt.target.dataset.name}"]`
-            )
-            .classList.add(`active`);
-
-          clearText();
-          history.pushState(
-            {
-              title: SearchFilter.getTitle(evt.target.dataset.name),
-              category: evt.target.dataset.name,
-              parent: "",
-              search: "",
-              filter: history.state?.filter,
-            },
-            SearchFilter.getTitle(evt.target.dataset.name),
-            evt.target.href
-          );
-
-          document.title = SearchFilter.getTitle(evt.target.dataset.name);
-          SearchFilter.filterSubcategory(evt.target.dataset.name);
-          SearchFilter.toggleSubcategory(true);
-          SearchFilter.updateHeader(evt.target.dataset.name, "");
-          SearchFilter.filterCategory(evt.target.dataset.name);
-        }
-      });
-    }
-
-    for (const subcategory of subcategories) {
-      subcategory.addEventListener(
-        "click",
-        (evt) => {
-          evt.stopImmediatePropagation();
-          if (evt.shiftKey || evt.metaKey || evt.ctrlKey || evt.altKey) {
-            return;
-          }
-
-          evt.preventDefault();
-
-          let href;
-
-          if (evt.target.dataset.name) {
-            clearText();
-            if (categoryTitle.value.trim() !== evt.target.dataset.name) {
-              categoryTitle.value = evt.target.dataset.name;
-              parentTitle.value = evt.target.dataset.parent;
-              href = evt.target.href;
-              SearchFilter.toggleSubcategory();
-              SearchFilter.highlightParent();
-            } else {
-              categoryTitle.value = evt.target.dataset.parent;
-              parentTitle.value = "";
-              href = document.querySelector(
-                `#multipage-nav a[data-name="${evt.target.dataset.parent}"]`
-              ).href;
-              SearchFilter.toggleSubcategory(true);
-            }
-
-            history.pushState(
-              {
-                title: SearchFilter.getTitle(evt.target.dataset.name),
-                category: categoryTitle.value.trim(),
-                parent: parentTitle.value.trim(),
-                search: "",
-                filter: history.state?.filter,
-              },
-              SearchFilter.getTitle(evt.target.dataset.name),
-              href
-            );
-
-            document.title = SearchFilter.getTitle(categoryTitle.value.trim());
-            SearchFilter.updateHeader(
-              categoryTitle.value.trim(),
-              parentTitle.value.trim()
-            );
-            SearchFilter.filterCategory(categoryTitle.value.trim());
-          }
-        },
-        true
-      );
-    }
-
-    document
-      .querySelector(`.go-back-to-all-link`)
-      .addEventListener("click", (evt) => {
-        evt.stopPropagation();
-
-        if (evt.shiftKey || evt.metaKey || evt.ctrlKey || evt.altKey) {
-          return;
-        }
-
-        evt.preventDefault();
-
-        clearText();
-        history.pushState(
-          {
-            title: SearchFilter.getTitle("None"),
-            category: "None",
-            parent: "",
-            search: "",
-            filter: history.state?.filter,
-          },
-          SearchFilter.getTitle(evt.target.dataset.name),
-          evt.target.href
-        );
-
-        document
-          .querySelector(`#multipage-nav a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector(`#pni-nav-mobile a.active`)
-          .classList.remove(`active`);
-
-        document
-          .querySelector(`#multipage-nav a[data-name="None"]`)
-          .classList.add(`active`);
-
-        document
-          .querySelector(`#pni-nav-mobile a[data-name="None"]`)
-          .classList.add(`active`);
-
-        SearchFilter.filterCategory("None");
-        parentTitle.value = "";
-      });
-
-    window.addEventListener(`popstate`, (event) => {
-      const { state } = event;
-      if (!state) return; // if it's a "real" back, we shouldn't need to do anything
-
-      const { title, category, parent } = state;
-      document.title = title;
-
-      if (!history.state?.search) {
-        SearchFilter.clearCategories();
-        categoryTitle.value = category;
-        parentTitle.value = parent;
-
-        searchBar.classList.remove(`has-content`);
-        searchInput.value = ``;
-
-        if (parent) {
-          SearchFilter.highlightParent();
-          SearchFilter.toggleSubcategory();
-        } else {
-          document
-            .querySelector(`#multipage-nav a.active`)
-            .classList.remove(`active`);
-
-          document
-            .querySelector(`#pni-nav-mobile a.active`)
-            .classList.remove(`active`);
-
-          document
-            .querySelector(`#multipage-nav a[data-name="${category}"]`)
-            .classList.add(`active`);
-
-          document
-            .querySelector(`#pni-nav-mobile a[data-name="${category}"]`)
-            .classList.add(`active`);
-
-          SearchFilter.toggleSubcategory(true);
-        }
-      } else {
-        SearchFilter.toggleSubcategory(true);
-        searchBar.classList.add(`has-content`);
-        searchInput.value = history.state?.search;
-        SearchFilter.filter(history.state?.search);
-      }
-
-      SearchFilter.filterCategory(category);
-      SearchFilter.filterSubcategory(parent || category);
-      SearchFilter.updateHeader(category, parent);
-
-      if (history.state?.filter) {
-        toggle.checked = history.state?.filter;
-
-        if (history.state?.filter) {
-          document.body.classList.add(`show-ding-only`);
-        } else {
-          document.body.classList.remove(`show-ding-only`);
-        }
-      }
-    });
-
-    history.replaceState(
-      {
-        title: SearchFilter.getTitle(categoryTitle.value.trim()),
-        category: categoryTitle.value.trim(),
-        parent: parentTitle.value.trim(),
-        search: history.state?.search ?? "",
-        filter: history.state?.filter,
-      },
-      SearchFilter.getTitle(categoryTitle.value.trim()),
-      location.href
-    );
-
-    if (history.state?.search) {
-      searchBar.classList.add(`has-content`);
-      searchInput.value = history.state?.search;
-      SearchFilter.filter(history.state?.search);
-    } else {
-      searchBar.classList.remove(`has-content`);
-      searchInput.value = ``;
-    }
-
-    if (history.state?.filter) {
-      toggle.checked = history.state?.filter;
-
-      if (history.state?.filter) {
-        document.body.classList.add(`show-ding-only`);
-      } else {
-        document.body.classList.remove(`show-ding-only`);
-      }
-    }
-
-    if (history.state?.parent && history.state?.category) {
-      document
-        .querySelector(
-          `a.subcategories[data-name="${history.state?.category}"]`
-        )
-        .scrollIntoView({
-          behavior: "smooth",
-          block: "nearest",
-          inline: "start",
-        });
-    }
+    new SearchFilterObject(SearchFilter);
   },
 
   clearCategories: () => {
